refactor: move commit and release structs in separate files
